unified-test-platform/requirements.txt
requests==2.28.1
selenium==4.7.2
appium-python-client==2.1.0
jinja2==3.1.2
pyyaml==6.0
pytest==7.2.0
pytest-html==3.2.0
allure-pytest==2.9.45
paramiko==2.12.0
docker==6.0.1
jenkinsapi==0.3.13
